fix(merge): include merge body in approval
When you merge with the dependabot option (alt-m), then we now include
the "@dependabot merge" body in the approval message instead of first
approving and then writing a comment.

Hopefully this solves a rare bug where dependabot doesn't actually merge
the PR and you have to manually comment with the same merge message again.