fix(commands): Change the "open in browser" keyboard from "b" to "o"
The "b" key is already defined as a keyboard shortcut in the
charmbracelet/bubbles library, which means whenever the "b" key is pressed, it
will both open the current item in a browser and navigate the list to the
previous page.

The "o" key is however not used for anything so doesn't collide with any other
actions.

Fixes https://github.com/einride/gh-dependabot/issues/89.