fix: add a rate limit to write operations
We have noticed while performing large amounts of merges during a very
short amount of time, that the Github API very quickly rate limits us,
leading to unmerged Pull Requests that amass over time.

We haven't been able to discover exactly which operations cause the rate
limiter to hit, but this Pull Request intends to improve on the
functionality.

Furthermore, we are proposing that we prevent the first quit while there
are operations in progress. The second quit will always go through. A
warning message is written if a quit is attempted while there are
ongoing operations.